requirements.txt
Flask==2.3.4
Flask-Login==0.6.2
Flask-Migrate==4.1.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
gunicorn==20.1.0
